+++ NEWS.md
@@ -1,3 +1,5 @@
+# hardhat (development version)
+
 # hardhat 1.3.0
 
 * New family of `spruce_*_multiple()` functions to support standardizing