+++ NEWS.md
@@ -1,5 +1,8 @@
 # hardhat (development version)
 
+* Changed an Rd name from `modeling-package` -> `modeling-usethis` at the
+  request of CRAN.
+
 # hardhat 1.3.0
 
 * New family of `spruce_*_multiple()` functions to support standardizing